Added functionality and conway ruleset
The game is now functional. You build a starting state, and then press "i" in order to start applying the rules. You can interrupt the game by pressing "i" again, which then allows you to edit the field. I need to iron out a bug where a glider will hit the bottom corner of the screen and appear at the top, and I'm working on it