Changed controls, added framerate/tickcount
You now start and stop life with the spacebar. The game is now locked to a certain refresh rate so won't chew *as much* of your CPU resources. I also removed some redundant/useless code, and changed the letters that represented living and dead cells to make them easier to read.